+++ MentalaisGidsAPI/Controllers/LietotajsController.cs
@@ -2,6 +2,7 @@
 using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
+using MentalaisGidsAPI.Domain;
 using MentalaisGidsAPI.Models;
 using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
